Socket Programming examples
the socket programming examples are as follows:
The four applications are
Echo client and server- single threaded
Simple Client server pair where client sends a message and server
receives and echos it back to the client.

echo server client-multithreaded:
Multithreaded version of client server pair.

A trivial date server and client, illustrating simple one-way
communication. The server sends data to the client only.

A two-player tic tac toe game, illustrating a server that needs to keep
track of the state of a game, and inform each client of it, so they can
each update their own displays.

A multi-user chat application, in which a server must broadcast
messages to all of its clients.